Issue #11 "Add NDjango view" dialog is now shown after adding new template from "Add / New Item -> NDjango simple page" Modified template to accepts parameter substitution in the template body

+++ NDjangoDesigner/Commands/AddViewDlg.cs
@@ -15,27 +15,88 @@ namespace NDjango.Designer.Commands
     {
         ViewWizard wizard = new ViewWizard();
         StringBuilder templateMem = new StringBuilder();
+        const string TEXT_NONE = "None";
+
+
         public AddViewDlg()
         {
             InitializeComponent();
             this.MinimumSize = this.Size; // minimum size is set to what is defined in designer
+            WriteItemDirectly = true;
         }
 
         public void FillDialogControls()
         {
             wizard.Update();
-            tbViewName.Text = wizard.GenerateName();
+            ViewName = wizard.GenerateName();
             FillModelList();
             FillAllTemplates();
             comboBaseTemplate.SelectedIndex = 0;//none value
             comboModel.SelectedIndex = 0;//none value
 
         }
-        
+
+        // PUBLIC PROPERTIES
+
+        /// <summary>
+        /// Always contains the pre-generated template text
+        /// </summary>
+        public string PreGeneratedTemplateText { get; set; }
+
+        /// <summary>
+        /// If set tu true then this form will directly generate the file with the specified name.
+        /// Otherwise this form will only set the pre-generated text to the "PreGeneratedTemplateText" field and return.
+        /// To ontain only pre-generated text without writing it to the file - set this value to false before showing the dialog
+        /// </summary>
+        public bool WriteItemDirectly { get; set; }
+
+        /// <summary>
+        /// Get or set the shown view name
+        /// </summary>
+        public string ViewName 
+        {
+            get { return tbViewName.Text; }
+            set { tbViewName.Text = value; }
+        }
+
+        public bool ViewNameEnabled
+        {
+            get { return tbViewName.Enabled; }
+            set { tbViewName.Enabled = value; }
+        }
+
+        public string SelectedModel
+        {
+            get { return (comboModel.SelectedItem == null || comboModel.SelectedItem.ToString() == "None") ? 
+                string.Empty : 
+                comboBaseTemplate.SelectedItem.ToString(); }
+        }
+
+        public string ModelToExtend
+        {
+            get { return (comboModel.SelectedItem == null || comboModel.SelectedItem.ToString() == "None") ? 
+                string.Empty : 
+                comboModel.SelectedItem.ToString(); }
+        }
+
+        // PRIVATE PROPERTIES
+
+        private bool IsInheritance 
+        {
+            get { return (comboBaseTemplate.SelectedItem != null && comboBaseTemplate.SelectedItem.ToString() != TEXT_NONE); } 
+        }
+
+        private bool IsViewModel 
+        {
+            get { return comboModel.SelectedItem != null && comboModel.SelectedItem.ToString() != TEXT_NONE; } 
+        }
+
+        // METHODS
+
         private void FillModelList()
         {
             comboModel.Items.Clear();
-            comboModel.Items.Add("None");
+            comboModel.Items.Add(TEXT_NONE);
             try
             {
                 List<Assembly> assmlist = wizard.GetReferences();
@@ -65,7 +126,7 @@ namespace NDjango.Designer.Commands
         private void FillAllTemplates()
         {
             comboBaseTemplate.Items.Clear();
-            comboBaseTemplate.Items.Add("None");
+            comboBaseTemplate.Items.Add(TEXT_NONE);
             IEnumerable<string> allTemplates = wizard.GetTemplates("");
             foreach (string item in allTemplates)
                 //if (!comboBaseTemplate.Items.Contains(item))
@@ -78,52 +139,71 @@ namespace NDjango.Designer.Commands
             templateMem.AppendLine("temp://{% extends \"" + comboBaseTemplate.SelectedItem + "\" %}");
             templateMem.AppendLine("{% block A %}");
             templateMem.AppendLine("{% endblock %}");
-
         }
         
         private void btnAdd_Click(object sender, EventArgs e)
         {
-            string itemName = tbViewName.Text + ".django";
-            wizard.RegisterInserted(comboBaseTemplate.SelectedItem.ToString());
-            string templateFile = Path.GetTempFileName();
-            StreamWriter sw = new StreamWriter(templateFile);
-            if (IsViewModel)
-                sw.WriteLine("{% model Model:" + comboModel.SelectedItem + " %}");
-            if (IsInheritance)
+            if (ViewName == string.Empty)
             {
-                sw.WriteLine("{% extends \"" + comboBaseTemplate.SelectedItem + "\" %}");
+                Misc.otherStuff.ShowErrorDialog("Error", "View Name can not be empty.");
+                return;
+            }
+
+            if (comboBaseTemplate.SelectedItem != null)
+                wizard.RegisterInserted(comboBaseTemplate.SelectedItem.ToString());
+
+            string itemName = ViewName + ".django";            
+            string templateFile = Path.GetTempFileName();
+            string PreGeneratedTemplateText = string.Empty;
+
+            // We need to generate these tags only if we're writing directly to the file. otherwise this will be generated on the fly
+            if (IsViewModel && WriteItemDirectly)
+                PreGeneratedTemplateText += "{% model Model:" + comboModel.SelectedItem + " %}\r\n";
+            
+            // We need to generate these tags only if we're writing directly to the file. otherwise this will be generated on the fly
+            if (IsInheritance && WriteItemDirectly)
+            {
+                PreGeneratedTemplateText += "{% extends \"" + comboBaseTemplate.SelectedItem + "\" %}\r\n";
                 if (checkedListBlocks.CheckedItems.Count > 0)
                 {
                     foreach (string name in checkedListBlocks.CheckedItems)
                     {
-                        sw.WriteLine("{% block " + name + " %}");
-                        sw.WriteLine("{% endblock " + name + " %}");
+                        PreGeneratedTemplateText += "{% block " + name + " %}\r\n";
+                        PreGeneratedTemplateText += "{% endblock " + name + " %}\r\n";
                     }
                 }
             }
-            sw.Close();
-            try
+
+            if (WriteItemDirectly)
             {
-                wizard.AddFromFile(templateFile, itemName);
-                this.Close();
+                using (StreamWriter sw = new StreamWriter(templateFile))
+                {
+                    sw.WriteLine(PreGeneratedTemplateText);
+                    sw.Close();
+                }
+
+                try
+                {
+                    wizard.AddFromFile(templateFile, itemName);
+                    this.Close();
+                }
+                catch (COMException ex)
+                {
+                    MessageBox.Show(ex.Message);
+                }
+                finally
+                {
+                    File.Delete(templateFile);
+                }
             }
-            catch (COMException ex)
-            {
-                MessageBox.Show(ex.Message);
-            }
-            finally
-            {
-                File.Delete(templateFile);
-            }
+            
         }
 
         private void btnCancel_Click(object sender, EventArgs e)
         {
             this.Close();
         }
-
-        private bool IsInheritance { get { return (comboBaseTemplate.SelectedItem != null && comboBaseTemplate.SelectedItem.ToString() != "None"); } }
-        private bool IsViewModel { get { return comboModel.SelectedItem != null && comboModel.SelectedItem.ToString() != "None"; } }
+        
         private void comboBaseTemplate_SelectedIndexChanged(object sender, EventArgs e)
         {
             if (IsInheritance)
